api/requirements.txt
fastapi
uvicorn[standard]
boto3
psycopg[binary]
sqlalchemy
alembic
python-jose[cryptography]
redis
rq
pydantic[email]
google-auth
requests